pyproject.toml
[tool.poetry]
name = "aiger_sat"
readme="README.md"
version = "0.1.0"
description = "Pythonic interface between AIGs and SAT solvers."
authors = ["Marcell Vazquez-Chanlatte <mvc@linux.com>"]
packages = [
    { include = "aiger_sat" },
]

[tool.poetry.dependencies]
python = "^3.6"
python-sat = {version = "^0.1.4.dev13",allows-prereleases = true}
py-aiger = "^3.3"
py-aiger-cnf = "^0.1.2"
attrs = "^19.1"

[tool.poetry.dev-dependencies]
codecov = "^2.0"
pytest = "^5.0"
pytest-cov = "^2.7"
pytest-flake8 = "^1.0"
pytest-sugar = "^0.9.2"
pytest-xdist = "^1.29"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
